Filter out special entry listed when the font cache is empty
When the folder is empty, the loop runs once with the text
of the wildcard, '.*.otf'. This results in an ineffective
symbolic link from a file named '*.otf' to a non-existing
file named '.*.otf' in the cache. While this has no averse
consequences, it is also an unintended side effect.

This script is similar to a version I was using in April 2021.
This is the end of the historical versions of the script.

+++ help-fontconfig-discover-hidden-fonts-for.sh
@@ -31,13 +31,16 @@ pwd
 echo 'Add symbolic links for current fonts...'
 for filePath in "$fontServiceCache"/.*.otf
 do
-  # extract file name
-  fileName="$(basename "$filePath")"
-  # remove initial '.'
-  linkName="${fileName#.}"
-  # create symbolic link
-  echo "$fileName -> $linkName"
-  ln -s "$filePath" "$linkName"
+  if test -f "$filePath"
+  then
+    # extract file name
+    fileName="$(basename "$filePath")"
+    # remove initial '.'
+    linkName="${fileName#.}"
+    # create symbolic link
+    echo "$fileName -> $linkName"
+    ln -s "$filePath" "$linkName"
+  fi
 done
 
 echo 'Refresh font cache...'
